+++ adapters/adf/adf_test.go
@@ -1,4 +1,4 @@
-package adformOpenRTB
+package adf
 
 import (
 	"testing"
@@ -9,12 +9,12 @@ import (
 )
 
 func TestJsonSamples(t *testing.T) {
-	bidder, buildErr := Builder(openrtb_ext.BidderAdformOpenRTB, config.Adapter{
+	bidder, buildErr := Builder(openrtb_ext.BidderAdf, config.Adapter{
 		Endpoint: "https://adx.adform.net/adx/openrtb"})
 
 	if buildErr != nil {
 		t.Fatalf("Builder returned unexpected error %v", buildErr)
 	}
 
-	adapterstest.RunJSONBidderTest(t, "adformopentrbtest", bidder)
+	adapterstest.RunJSONBidderTest(t, "adftest", bidder)
 }
