+++ adapters/adf/adftest/exemplary/multi-native.json
