+++ adapters/adf/adftest/exemplary/single-native.json
