+++ adapters/adf/adftest/supplemental/bad-request.json
