+++ exchange/adapter_builders.go
@@ -4,8 +4,8 @@ import (
 	"github.com/prebid/prebid-server/adapters"
 	ttx "github.com/prebid/prebid-server/adapters/33across"
 	"github.com/prebid/prebid-server/adapters/acuityads"
+	"github.com/prebid/prebid-server/adapters/adf"
 	"github.com/prebid/prebid-server/adapters/adform"
-	"github.com/prebid/prebid-server/adapters/adformOpenRTB"
 	"github.com/prebid/prebid-server/adapters/adgeneration"
 	"github.com/prebid/prebid-server/adapters/adhese"
 	"github.com/prebid/prebid-server/adapters/adkernel"
@@ -117,8 +117,8 @@ func newAdapterBuilders() map[openrtb_ext.BidderName]adapters.Builder {
 	return map[openrtb_ext.BidderName]adapters.Builder{
 		openrtb_ext.Bidder33Across:         ttx.Builder,
 		openrtb_ext.BidderAcuityAds:        acuityads.Builder,
+		openrtb_ext.BidderAdf:              adf.Builder,
 		openrtb_ext.BidderAdform:           adform.Builder,
-		openrtb_ext.BidderAdformOpenRTB:    adformOpenRTB.Builder,
 		openrtb_ext.BidderAdgeneration:     adgeneration.Builder,
 		openrtb_ext.BidderAdhese:           adhese.Builder,
 		openrtb_ext.BidderAdkernel:         adkernel.Builder,
