+++ openrtb_ext/bidders_validate_test.go
@@ -50,7 +50,7 @@ func TestBidderUniquenessGatekeeping(t *testing.T) {
 	// - Exclude duplicates of adapters for the same bidder, as it's unlikely a publisher will use both.
 	var bidders []string
 	for _, bidder := range CoreBidderNames() {
-		if bidder != BidderTripleliftNative && bidder != BidderAdkernelAdn && bidder != BidderAdformOpenRTB {
+		if bidder != BidderTripleliftNative && bidder != BidderAdkernelAdn {
 			bidders = append(bidders, string(bidder))
 		}
 	}
