+++ static/bidder-info/adf.yaml
