+++ static/bidder-params/adf.json
@@ -1,7 +1,7 @@
 {
   "$schema": "http://json-schema.org/draft-04/schema#",
-  "title": "AdformOpenRTB Adapter Params",
-  "description": "A schema which validates params accepted by the AdformOpenRTB adapter",
+  "title": "Adf Adapter Params",
+  "description": "A schema which validates params accepted by the adf adapter",
   "type": "object",
   "properties": {
     "mid": {
